+++ NEWS.md
@@ -26,8 +26,8 @@ objects.
 version of RNifti.
 
 # spant 1.18.0
-* Better behavior when using stackplot with single spectra datasets.
-* Better behavior when using image with single spectra datasets.
+* Better behaviour when using stackplot with single spectra datasets.
+* Better behaviour when using image with single spectra datasets.
 * Bug fix for reading MEGA-PRESS data in TWIX format.
 * Better error checking for lists of spectra.
 * Bug fix for calculating NAA linewidth from MEGA-PRESS data.